Add UI and add calculator functions
This commit creates a basic UI which contains buttons 0-9 as well as
4 operators +, -, *, /. There are also two additional buttons =
which will perform the calculation and Clr which will reset the
calculator back to 0. As of this commit the buttons are not yet
functional. This commit adds five functions to the calculator in
script.js add(), subtract(), multiply(), divide() and operate().
The operate() function takes the input of an operator and two numbers
and will perform the calculations of the operator and return the total.