Add performCalculations() function, allow multiple operations
This commit creates the performCalculations() function and allows for
the user to string together multiple calculations at once.
performCalculations() works by using commas around operators in
displayValue which we can use to split the string into an array
around the operators. performCalculations() will then check to see
if there are any operators left in the displayValue. This will let the
function know before displaying the total if there are any more
calculations to perform. This commit also changes the functionality of
the = button. Instead of displaying = in the calculator's display the
display will now be cleared and showed only the total of the
calculations.