Add decimal button, add decimal error, change parseInt() to parseFloat()
This commit adds decimal place support to the calculator. If two
decimals are entered for one number the user will see an error
that multiple decimals were entered for a single number. parseInt() is
used in the performCalculations() function to convert the array element
to a number in the previous commit. It now has been changed to
parseFloat(). Without parseFloat() all numbers will become rounded to
the nearest integer.