Add negative number support
This commit adds support for negative numbers. In the
performCalculations() function two checks were added. The first check is
to see if the first number is a negative. The second check is to
identify any negative numbers in the displayValueArray. This
determination is made if there are two subtraction signs in adjacent
array elementsa row or if the first value of displayValueArray is a
subtraction sign.