Fix bug when only one number is entered
This commit fixes a bug that occurs if the user only enters one number
into the calculator and hits enter. Prior to this commit the total
will always equal zero in this case. After this commit the display now
equals the only number on the calculator's display. This was
accomplished through the use of a global variable calculatedOnce which
will allow the script to determine if the calculator performed one
operation after the user hit equals. This is important because it allows
us to determine if the user entered one number only.